third_party/blink/web_tests/wpt_internal/bluetooth/characteristic/notifications/characteristic-does-not-support-notifications.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattService => gattService.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('body_sensor_location'))
    .then(characteristic => assert_promise_rejects_with_message(
      characteristic.startNotifications(),
      new DOMException('GATT Error: Not supported.',
                       'NotSupportedError'),
      'Characteristic does not support notifications.'));
}, 'Characteristic does not support notifications. Reject with ' +
   'NotSupportedError.');
</script>
